+++ project/Deps.scala
@@ -1,17 +1,18 @@
 import sbt._
 
 object Deps {
-  lazy val http4sVersion             = "1.0.0-M40"
+  lazy val http4sVersion             = "1.0.0-M41"
   lazy val log4catsVersion           = "2.4.0"
-  lazy val scalatestVersion          = "3.2.17"
-  lazy val circeVersion              = "0.14.6"
+  lazy val scalatestVersion          = "3.2.18"
+  lazy val circeVersion              = "0.14.7"
   lazy val circeGenericExtrasVersion = "0.14.3"
   lazy val circeYamlVersion          = "0.15.1"
-  lazy val fs2Version                = "3.2.2"
-  lazy val pulsarVersion             = "3.1.2"
-  lazy val luceneVersion             = "9.9.1"
-  lazy val awsVersion                = "2.23.3"
+  lazy val fs2Version                = "3.10.2"
+  lazy val pulsarVersion             = "3.2.2"
+  lazy val luceneVersion             = "9.10.0"
+  lazy val awsVersion                = "2.25.15"
   lazy val prometheusVersion         = "0.16.0"
+  lazy val djlVersion = "0.27.0"
 
   val httpsDeps = Seq(
     "org.http4s" %% "http4s-dsl"          % http4sVersion,