+++ src/main/scala/ai/metarank/source/format/JsonFormat.scala
@@ -27,7 +27,7 @@ object JsonFormat extends SourceFormat with Logging {
           case Left(value) =>
             logger.error(s"Cannot parse json input: '${new String(next.toArray)}'", value)
             throw value
-          case Right(value) => (parser, Chunk.seq(value))
+          case Right(value) => (parser, Chunk.from(value))
         }
       })
       .evalMapChunk(json =>